backend/requirements.txt
fastapi
python-multipart
uvicorn
pydantic
llama-index
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface
llama-index-llms-litellm
langchain
langchain-community
langchain-huggingface
langchain-google-genai
langchain_pinecone
pinecone-client
qdrant-client
sentence-transformers
pypdf
python-dotenv
litellm
redis
celery
coqui-tts
tavily-python
langchain-tavily
langchain-ollama
langgraph
llama-index-llms-langchain
llama-index-llms-gemini
